\name{build.colony.input}
\alias{build.colony.input}

\title{A wizard-like function to help build Colony2 input files.}

\description{
This function prompts the user to provide information about their system in order to correctly build and format a Colony2 input file.
}

\usage{
build.colony.input(wd=getwd(),name="Colony2.DAT")

}

\arguments{
  \item{wd}{The directory where the generated file will be placed. The default is the current working directory.}
  \item{name}{The desired filename for the Colony input file (the default is Colony2.DAT).}
 }

\details{
  
}

\value{
A text file is produced.
}

\references{}
\author{Owen R. Jones}
\note{}
\seealso{\code{\link{run.colony}}}
\examples{

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }
