% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.colony.superauto.R
\name{build.colony.superauto}
\alias{build.colony.superauto}
\title{This function builds an automatic .DAT file with parameters that I (Jenna) am
using to test some simulated data. It should NOT be used to build .DAT files
for other projects, as it will NOT prompt for all of the options.
This is purely for my convenience and to speed up testing!}
\usage{
build.colony.superauto(
  wd = getwd(),
  name,
  datasetname,
  delim = "",
  sample_size,
  num_loci,
  error_rates_path,
  genotypes_path
)
}
\arguments{
\item{wd}{The directory where the generated file will be placed. The default
is the current working directory.}

\item{name}{The desired filename for the Colony input file (the default is
Colony2.DAT).}

\item{delim}{What is the delimiter for the input files (default is that they
are whitespace delimited).}
}
\value{
A text file is produced. This file can be used by Colony2 as an
input file.
}
\description{
This function builds an automatic .DAT file with parameters that I (Jenna) am
using to test some simulated data. It should NOT be used to build .DAT files
for other projects, as it will NOT prompt for all of the options.
This is purely for my convenience and to speed up testing!
}
\references{
Wang, J. (2004) Sibship reconstruction from genetic data with
typing errors.  Genetics 166: 1963-1979.
}
\seealso{
\code{\link{run.colony}}
}
\author{
Owen R. Jones
}
\keyword{manip}
