\name{rcolony-package}
\alias{rcolony-package}
\alias{rcolony}
\docType{package}
\title{
Tools for implementation and visualisation of Colony and its outputs.}
\description{
The RColony package provides tools for the implementation of the Colony pedigree analysis software. In addition it provides tools for visualising, and summarising the results of the Colony analyses.}
\details{
\tabular{ll}{
Package: \tab rcolony\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-04-16\cr
License: \tab GNU General Public License (GPL)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
