% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.colony.R
\name{run.colony}
\alias{run.colony}
\title{A function to run Colony2 from within the R environment.}
\usage{
run.colony(colonyexecpath = "prompt", datfilepath = "prompt", wait = TRUE)
}
\arguments{
\item{colonyexecpath}{The path to the Colony2 executable. If none is given, and the function defaults to \"prompt\", then you will be prompted to click on the file.}

\item{datfilepath}{The path to the DAT file that you wish to run. If none is given, and the function defaults to \"prompt\", then you will be prompted to click on the file.}

\item{wait}{Should R wait for Colony2 to be run by the system or not? TRUE/FALSE}
}
\value{
return value (0 if success)
}
\description{
This function used \code{system2} to call Colony2 from within R.
}
\references{
Wang, J. (2004) Sibship reconstruction from genetic data with typing errors.  Genetics 166: 1963-1979.
}
\seealso{
\code{\link{monitor.colony}}
}
\author{
Timothee Flutre

Owen R. Jones
}
\keyword{IO}
