\name{testdata}
\alias{testdata}

\title{Example dataset for the rcolony package}
\description{
This is an example dataset, as would be obtained using \code{get.colony.data}}
\usage{
data(testdata)
}
\format{
The datafile contains a list (\code{testdata}) as generated with  \code{get.colony.data}. There are nine variables: 
\describe{
\item{fathers}{Paternity data.}
\item{mothers}{Maternity data.}
\item{offspring}{Offspring data.}
\item{maternity}{Maternity assignments.}
\item{paternity}{Paternity assignments.}
\item{pairwise.maternity}{Pairwise maternity assignments.}
\item{pairwise.paternity}{Pairwise paternity assignments.}
\item{sibs}{Sibship (half and full) assignments.}
\item{pairwise.sibs}{Pairwise sibship (half and full) assignments.}


}
}


\references{}
\seealso{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}

