\name{rcolony-package}
\alias{rcolony-package}
\alias{rcolony}

\alias{}
\docType{package}
\title{
Tools for implementation and visualisation of Colony and its outputs.
}
\description{The \pkg{rcolony} provides tools for the implementation of the Colony pedigree analysis software. In addition it provides tools for visualising, and summarising the results of the Colony analyses. The package is designed to work with Colony v.2.0 and some functions may not work with other versions.
}
\details{
\tabular{ll}{
Package: \tab rcolony\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-12-12\cr
License: \tab GNU GPL\cr
}

}
\author{
Owen R. Jones
Maintainer: Owen R. Jones <jones@demogr.mpg.de>
}
\references{
Jones, O. and Wang, J. (2009) COLONY: a program for parentage and sibship inference from multilocus genotype data. Molecular Ecology Resources 10: 551 - 555.

Wang, J. (2004) Sibship reconstruction from genetic data with typing errors. Genetics 166: 1963 - 1979.
}

\keyword{package}
\seealso{}
\examples{}
