\name{rcolony-package}
\alias{rcolony-package}
\alias{rcolony}
\docType{package}
\title{A package for producing input files, running, and plotting outputs from Colony2.}
\description{The RColony package provides tools for the implementation of the Colony pedigree analysis software. In addition it provides tools for visualising, and summarising the results of the Colony analyses.}

\details{

\tabular{ll}{
Package: \tab rcolony\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-16\cr
License: \tab GNU General Public License (GPL)\cr
LazyLoad: \tab yes\cr
}
}

\author{Owen R. Jones <owen.jones@ioz.ac.uk>}

\references{
\url{http://www.zoo.cam.ac.uk/ioz/software.htm#COLONY}
Wang, J. (2004) Sibship reconstruction from genetic data with typing errors.  Genetics 166: 1963-1979.}

\keyword{package}
\seealso{}
\examples{}
